Story Mode Enabled
After a ton of digging, I found where the demo mode is flagged at the start of the game. I also added some minor comment notes to some other stuff I tried on this page. There were a ton of other attempts made but noting that I was able to understand so they were all reverted before this commit since I added nothing to the knowledge pool sadly. But this is a major first step. Now to figure out how to adjust the rival's teams so that it is actually able to be won possibly.