Rival Battle Implementation (Mostly)
The Rival battle has now been implemented into the game as it should properly. The only problem is that the intro text when entering the battle is grammatically incorrect due to an issue I am having while editing the parties.asm file. It currently shows the text of "<rival's name>'s has started a battle!" but I will attempt to fix this later.

I also updated the README to show these feature changes.