Barricades Removed (& Some Translation/Definition Comments)
Modified PrinceRoute.blk and SilentHill.blk using a tool called polishmap to remove the barricades that were implemented for the demo to prevent players from exploring where they weren't supposed to.

Updated the README to include these changes.

Made comments in SilentHill.asm regarding how the object script works that I learned while attempting to start working on a temporary blockade for the East exit of town like what Blue does for the West exit to prevent soft locks from entering a battle without a Pokemon in your party.

Added comments to main_menu.asm that show translations for some of the text in the game. This does NOT effect the language of the actual gameplay..