Start Menu & Debug Menu Update (And Minor Translation Comments)
Changed the start menu from the one used by the demo to a proper full menu. This needs more research as save does not seem to either work or it just doesn't trigger the continue option on the title screen. Research also needs to be done on the story mode flags to properly modify the start menu that is loaded as they don't seem to be either triggered or at least called.

Also added the debug field menu to the game so that the user can heal their team and even access some other features like the unfinished storage boxes in the PC and some Mart features.

Added a handful of translations to the notes within the code as well. This does not effect the gameplay at all.