Temporary Shiny Hunter's Patch
This includes changes to SilentHillLabFront to change the 6 Pokeballs given to you buy Oak's granddaughter to 999. This causes a visual bug but seems to be stable from my limited testing. I did this to allow shiny hunters to play the game more openly since I was able to get what I believe to be 2 shiny Pokemon to spawn and thought this would be fun for people while I worked on more features.

I also updated pokecenter_pc to have a very jank but semi-functional PC menu now. This functions only within the debug menu for the time being while I finish sorting out how to get it working the right way. This also only adds 20 more spaces since the new box system was not completed by GameFreak at this point. But hey, at least you can catch 26 things instead of just 5 since you couldn't release your starter before. Patch will be released for this once I can either fix the PC fully or I decide that this is good enough for people for now. We will see in a day or two how I feel with this part.