V9 Readme Updates & Minor Revert
Reverting the changes to the makefile since it didn't actually fix my issue when building. I think the issue is the way I setup things on my laptop sadly.

Updated the README to show the changes from the V9 patch as well as adding in a list for known bugs to make it more clear what is actually broken.

Updating .gitignore to try and stop .DS_Store from being uploaded.