Tileset Names Connected & Map Collision Correction
The main changes of importance are those in constants/tileset_constants.asm and gfx/gfx.asm as the comments link the 2 files together better. The tilesets use different names from one another so I matched them up in the comments for better legibility.

I have also begun working on correcting the collision data for the other maps with this update. Since this is so manual, I have only started Old City as a proof of concept. More work will be done to slowly correct all maps that use the old and incorrect collision data.