Oak's Lab Lockout Revert & More Collision Fixes
Seemingly, due to messing with the collision data, I had to revert the changes to the demo lockout of Oak's Lab. For some reason after working with this data, it seems to now cause a crash upon exiting his lab after defeating your rival. It is also possible that this is due to me force removing the demo end dialogue with the rival at the end of Route 1. I did not do much testing on this as it is not a major setback for the time being and the collision data is slightly more important.

I finished all of the Old City tileset collision data as well as the Route Gate tileset collision data. This should now resolve the route to the west of Old City between it and the town called "West". It should also theoretically fix all Route Gates within the game as far as I am currently aware from limited testing.