West, Baadon, & Connecting Routes Collision Data
Adding collision for West (city), Baadon (city) and the connecting routes that share the same tilesets. This opens up a very large section of the centre of the map as now properly explorable. This includes being able to battle and catch all the corresponding wild Pokemon on each of these routes.