Collision New & Corrected
I have corrected a few of the warp collisions. Namely the stair warps and countertops in the Route Gates as to match what is seen in the officially updated tileset collision data from TPC.

I also have updated the collision data for the following tilesets: Marts (possibly unfinished), Fonto, & Blue Forest. This includes the surrounding routes of each area that share the corresponding tilesets.

This should bring the game up to roughly ~80% playable as far as map collision data is concerned.