All Needed Collision Data Fixed
This includes collision data updates for all the used tilesets in the game currently. The overworld is now technically 100% complete.

The only ones currently not updated are Cave, Power Plant, SS Anne, and SS Anne Dock. I will update all of these later as it appears they were intended to be used in the future but just not as of this build.

The tilesets that were completely updated for this commit are High Tech, Kanto, South, Rocket House, Office, and Ruins of Alph.

I made some corrections to the Gym tileset as they appear to have been mid-conversion by TPC.

I started working on the tilesets for Cave and Power Plant but realised that even though they have maps in this build of the game, they were not actually linked to any parts of the map so they are technically unused,

I also removed the @ from the code in the Rival starter parties. This doesn't effect anything in the gameplay, just removes an unnecessary item from the code.