Clean Setup Instructions (Final Update?)
Updating the "macOS Mojave (10.14.6) Clean Setup Instructions" section with what I hope to be the final major changes.

I am going to wipe the external SSD I did all of this on to start from scratch and follow my instructions to see if there is anything that I missed that would cause this not to be a perfect set of instructions for at least my machine.

If this works, then this should hopefully be at least a very close guideline for other similar machines to be able to follow and use themselves.