Final Update to the Clean Install Instructions Readme
As of the moment, I do not have it in me to attempt to wipe the external boot drive that I use for the fourth time this month alone. So I believe this is the final update to the instructions.

But I followed the instructions the whole way through before finding out I forgot to install Python 3. So this has been added to the instructions and theoretically should now function properly.

I am going to compare some things against my internal boot drive and make sure that there are no immediately noticeable issues with the install instructions maybe not compiling the ROM correctly or something. If there is a problem with that, then I will try to troubleshoot the issue and consider one more reset of the external boot drive to see if the instructions can be perfected.