Fixing Rival Battle Bug
Looks like in a previous commit I had been attempting to fix the grammatical error in the Rival battle and forgot to revert the changes. This caused the party to revert to the bugged format. I have now fixed this issue and hopefully you should be able to compile the latest version again properly.