Fixed Old City Mart Map
For some reason the Old City Mart's map was changed to be 6 tiles wide instead of 8 tiles wide like all the other Marts in the game. This caused the map layout to be corrupted as it tried to fit the full seemingly standard map layout into the smaller space. This has been remapped to look similar to the larger maps but shrunk down to the best of my ability.