Debug PC Fix & Player Room Full PC Access
This code fixes the debug menu PC to function how it actually should as if it were a real PC or in Field Debug Mode. These fixes inadvertently allowed the user to have full PC access in the Player's bedroom as if it were a Pokemon Center PC. So have fun with this.

I am still attempting to see if I can get the Pokemon Center PC functional, but for now, this is a pretty major step. Especially since I have the Pokemon Center PC code commented out due to a value overflow. You can technically enable it if you want, but it will cause the Lab events to be out of sync which prevents you from being able to access the starter Pokemon.