Bag Expansions & Debug Cut (Plus Readme Changes)
This makes items like the Balls actually go into the Ball Holder item and changes Nanami's code to actually give it to the player liker her Dialogue suggests she should be. Also adjusts the 8 items that the player is given by Nanami. Might try to even use the better Pack with Pockets from debug eventually but not sure how plausible based on supposed craash bugs in its code.

This also adds a new debug option of Cut to the bottom of the second page. This works seemingly everywhere except Old which might be due to how the map data is done. Might try to fix that later when I remove a lot of other map locks throughout the world that I missed in the past.

README updated to match bugs formatting found in upcoming V14 release notes as well as some other minor tweaks for legibility.

Only really useful for me, but I also added a bunch of notes to items and such to help myself in future edits/fixes.