eisnerguy1's Shiny Pinkmon Colour Palette Fix
A fix for the shiny pinkmon sprite palette that I didn't even realise was broken. This was taken with permission from eisnerguy1's work on the project which can be found at the link below. Will be trying to incorporate more of their work into this project if there are any other patches that haven't already been done.

https://github.com/eisnerguy1/pokegold-spaceworld-restore

(Also some other minor file changes related to me trying to fix the shiny sparkle effect that I commented back out because it doesn't work yet.)