eisnerguy1 PokeMart Fix & Minor Revert
I apparently missed the marts in High Tech and North that had the same size issue as Old Town. I reverted the map file for Old Town to the one in the base game and used the size adjustment that eisnerguy1 used in their patch to correct all 3 of these marts.

Also some new comments on things that I figured out while trying to test other feature updates.